:recycle: Create an app module.
